+++ src/HealthcareApis/HealthcareApis/HealthcareApis/SetAzureRmHealthcareApisService.cs
@@ -24,6 +24,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Management.Automation;
 
+
 namespace Microsoft.Azure.Commands.HealthcareApis.Commands
 {
     [Cmdlet(VerbsCommon.Set, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "HealthcareApisService", DefaultParameterSetName = ServiceNameParameterSet, SupportsShouldProcess = true), OutputType(typeof(PSHealthcareApisService))]
@@ -33,87 +34,193 @@ namespace Microsoft.Azure.Commands.HealthcareApis.Commands
         protected const string ResourceIdParameterSet = "ResourceIdParameterSet";
         protected const string InputObjectParameterSet = "InputObjectParameterSet";
 
-        [Parameter(Mandatory = true, ParameterSetName = ServiceNameParameterSet, HelpMessage = "HealthcareApis Service Name.")]
+        [Parameter(
+            Mandatory = true,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "HealthcareApis Service Name.")]
         [ValidateNotNullOrEmpty]
         [ValidateLength(2, 64)]
         public string Name { get; set; }
 
-        [Parameter(Mandatory = true, ParameterSetName = ServiceNameParameterSet, HelpMessage = "HealthcareApis Service Resource Group Name.")]
+        [Parameter(
+            Mandatory = true,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "HealthcareApis Service Resource Group Name.")]
         [ResourceGroupCompleter]
         [ValidateNotNullOrEmpty]
         public string ResourceGroupName { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = ServiceNameParameterSet, HelpMessage = "HealthcareApis FhirService CosmosOfferThroughput.")]
-        [Parameter(Mandatory = false, ParameterSetName = ResourceIdParameterSet, HelpMessage = "HealthcareApis FhirService CosmosOfferThroughput.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "HealthcareApis FhirService CosmosOfferThroughput.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "HealthcareApis FhirService CosmosOfferThroughput.")]
         [ValidateNotNullOrEmpty]
         public int? CosmosOfferThroughput { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = ServiceNameParameterSet, HelpMessage = "HealthcareApis FhirService Authority.")]
-        [Parameter(Mandatory = false, ParameterSetName = ResourceIdParameterSet, HelpMessage = "HealthcareApis FhirService Authority.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "HealthcareApis Fhir Service CosmosKeyVaultKeyUri. The URI of the customer-managed key for the backing database.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "HealthcareApis Fhir Service CosmosKeyVaultKeyUri. The URI of the customer-managed key for the backing database.")]
+        [ValidateNotNullOrEmpty]
+        public string CosmosKeyVaultKeyUri { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "HealthcareApis FhirService Authority.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "HealthcareApis FhirService Authority.")]
         [ValidateNotNullOrEmpty]
         public string Authority { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = ServiceNameParameterSet, HelpMessage = "HealthcareApis FhirService Audience.")]
-        [Parameter(Mandatory = false, ParameterSetName = ResourceIdParameterSet, HelpMessage = "HealthcareApis FhirService Audience.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "HealthcareApis FhirService Audience.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "HealthcareApis FhirService Audience.")]
         [ValidateNotNullOrEmpty]
         public string Audience { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = ServiceNameParameterSet, HelpMessage = "HealthcareApis FhirService EnableSmartProxy.")]
-        [Parameter(Mandatory = false, ParameterSetName = ResourceIdParameterSet, HelpMessage = "SmartProxyEnabled.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "HealthcareApis FhirService EnableSmartProxy.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "SmartProxyEnabled.")]
         [ValidateNotNullOrEmpty]
         public SwitchParameter EnableSmartProxy { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = ServiceNameParameterSet, HelpMessage = "HealthcareApis FhirService DisableSmartProxy.")]
-        [Parameter(Mandatory = false, ParameterSetName = ResourceIdParameterSet, HelpMessage = "HealthcareApis FhirService DisableSmartProxy.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "HealthcareApis FhirService DisableSmartProxy.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "HealthcareApis FhirService DisableSmartProxy.")]
         [ValidateNotNullOrEmpty]
         public SwitchParameter DisableSmartProxy { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = ServiceNameParameterSet, HelpMessage = "HealthcareApis FhirService List of Cors Origins. Specify URLs of origin sites that can access this API, or use \" * \" to allow access from any site.")]
-        [Parameter(Mandatory = false, ParameterSetName = ResourceIdParameterSet, HelpMessage = "HealthcareApis FhirService List of Cors Origins. Specify URLs of origin sites that can access this API, or use \" * \" to allow access from any site.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "HealthcareApis FhirService List of Cors Origins. Specify URLs of origin sites that can access this API, or use \" * \" to allow access from any site.")]
+        [Parameter(Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "HealthcareApis FhirService List of Cors Origins. Specify URLs of origin sites that can access this API, or use \" * \" to allow access from any site.")]
         [ValidateNotNullOrEmpty]
         public string[] CorsOrigin { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = ServiceNameParameterSet, HelpMessage = "HealthcareApis FhirService List of Cors Headers. Specify HTTP headers which can be used during the request. Use \" * \" for any header.")]
-        [Parameter(Mandatory = false, ParameterSetName = ResourceIdParameterSet, HelpMessage = "HealthcareApis FhirService List of Cors Headers. Specify HTTP headers which can be used during the request. Use \" * \" for any header.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "HealthcareApis FhirService List of Cors Headers. Specify HTTP headers which can be used during the request. Use \" * \" for any header.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "HealthcareApis FhirService List of Cors Headers. Specify HTTP headers which can be used during the request. Use \" * \" for any header.")]
         [ValidateNotNullOrEmpty]
         public string[] CorsHeader { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = ServiceNameParameterSet, HelpMessage = "HealthcareApis FhirService List of Cors Methods.")]
-        [Parameter(Mandatory = false, ParameterSetName = ResourceIdParameterSet, HelpMessage = "HealthcareApis FhirService List of Cors Methods.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "HealthcareApis FhirService List of Cors Methods.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "HealthcareApis FhirService List of Cors Methods.")]
         [ValidateNotNullOrEmpty]
         public string[] CorsMethod { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = ServiceNameParameterSet, HelpMessage = "HealthcareApis FhirService Cors Max Age. Specify how long a result from a request can be cached in seconds. Example: 600 means 10 minutes.")]
-        [Parameter(Mandatory = false, ParameterSetName = ResourceIdParameterSet, HelpMessage = "HealthcareApis FhirService Cors Max Age. Specify how long a result from a request can be cached in seconds. Example: 600 means 10 minutes.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "HealthcareApis FhirService Cors Max Age. Specify how long a result from a request can be cached in seconds. Example: 600 means 10 minutes.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "HealthcareApis FhirService Cors Max Age. Specify how long a result from a request can be cached in seconds. Example: 600 means 10 minutes.")]
         [ValidateNotNullOrEmpty]
         public int? CorsMaxAge { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = ServiceNameParameterSet, HelpMessage = "HealthcareApis FhirService AllowCorsCredentials.")]
-        [Parameter(Mandatory = false, ParameterSetName = ResourceIdParameterSet, HelpMessage = "HealthcareApis FhirService AllowCorsCredentials.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "HealthcareApis FhirService AllowCorsCredentials.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "HealthcareApis FhirService AllowCorsCredentials.")]
         [ValidateNotNullOrEmpty]
         public SwitchParameter AllowCorsCredential { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = ServiceNameParameterSet, HelpMessage = "HealthcareApis FhirService CorsCredentials Not Allowed.")]
-        [Parameter(Mandatory = false, ParameterSetName = ResourceIdParameterSet, HelpMessage = "HealthcareApis FhirService CorsCredentials Not Allowed.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "HealthcareApis FhirService CorsCredentials Not Allowed.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "HealthcareApis FhirService CorsCredentials Not Allowed.")]
         [ValidateNotNullOrEmpty]
         public SwitchParameter DisableCorsCredential { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = ServiceNameParameterSet, HelpMessage = "HealthcareApis Fhir Service Export Storage Account Name.")]
-        [Parameter(Mandatory = false, ParameterSetName = ResourceIdParameterSet, HelpMessage = "HealthcareApis Fhir Service Export Storage Account Name.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "HealthcareApis Fhir Service Export Storage Account Name.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "HealthcareApis Fhir Service Export Storage Account Name.")]
         [ValidateNotNullOrEmpty]
         public string ExportStorageAccountName { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = ServiceNameParameterSet, HelpMessage = "List of Access Policy Object IDs.")]
-        [Parameter(Mandatory = false, ParameterSetName = ResourceIdParameterSet, HelpMessage = "List of Access Policy Object IDs.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "List of Access Policy Object IDs.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "List of Access Policy Object IDs.")]
         [ValidateNotNullOrEmpty]
         public string[] AccessPolicyObjectId { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = ServiceNameParameterSet, HelpMessage = "Enable Managed Identity.")]
-        [Parameter(Mandatory = false, ParameterSetName = ResourceIdParameterSet, HelpMessage = "Enable Managed Identity.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "Enable Managed Identity.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "Enable Managed Identity.")]
         public SwitchParameter EnableManagedIdentity { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = ServiceNameParameterSet, HelpMessage = "Disable Managed Identity.")]
-        [Parameter(Mandatory = false, ParameterSetName = ResourceIdParameterSet, HelpMessage = "Disable Managed Identity.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "Disable Managed Identity.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "Disable Managed Identity.")]
         public SwitchParameter DisableManagedIdentity { get; set; }
 
         [Parameter(
@@ -129,23 +236,39 @@ namespace Microsoft.Azure.Commands.HealthcareApis.Commands
         [AllowEmptyCollection]
         public Hashtable Tag { get; set; }
 
-        [Parameter(Mandatory = true, ParameterSetName = InputObjectParameterSet, HelpMessage = "HealthcareApis fhir service piped from Get-AzHealthcareApisFhirService.", ValueFromPipeline = true)]
+        [Parameter(
+            Mandatory = true,
+            ParameterSetName = InputObjectParameterSet,
+            HelpMessage = "HealthcareApis fhir service piped from Get-AzHealthcareApisFhirService.", ValueFromPipeline = true)]
         public PSHealthcareApisService InputObject { get; set; }
 
 
-        [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, ParameterSetName = ResourceIdParameterSet, HelpMessage = "HealthcareApis Fhir Service ResourceId.")]
+        [Parameter(
+            Mandatory = true,
+            ValueFromPipelineByPropertyName = true,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "HealthcareApis Fhir Service ResourceId.")]
         [ResourceIdCompleter("Microsoft.HealthcareApis/services")]
         [ValidateNotNullOrEmpty]
         public string ResourceId { get; set; }
 
         [Parameter(
-         Mandatory = false,
-         HelpMessage = "Run cmdlet as a job in the background.")]
+            Mandatory = false,
+            HelpMessage = "Run cmdlet as a job in the background.")]
         public SwitchParameter AsJob { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = ServiceNameParameterSet, HelpMessage = "The network access type for Fhir service. Commonly `Enabled` or `Disabled`.")]
-        [Parameter(Mandatory = false, ParameterSetName = ResourceIdParameterSet, HelpMessage = "The network access type for Fhir service. Commonly `Enabled` or `Disabled`.")]
-        [ValidateSet("Enabled", "Disabled", IgnoreCase = true)]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "The network access type for Fhir service. Commonly `Enabled` or `Disabled`.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "The network access type for Fhir service. Commonly `Enabled` or `Disabled`.")]
+        [ValidateSet(
+            "Enabled",
+            "Disabled",
+            IgnoreCase = true)]
         public string PublicNetworkAccess { get; set; }
 
         public override void ExecuteCmdlet()
@@ -163,7 +286,7 @@ namespace Microsoft.Azure.Commands.HealthcareApis.Commands
                             {
                                 string rgName = null;
                                 string name = null;
-                                
+
                                 if (ParameterSetName.Equals(ResourceIdParameterSet))
                                 {
                                     ValidateAndExtractName(this.ResourceId, out rgName, out name);
@@ -272,6 +395,7 @@ namespace Microsoft.Azure.Commands.HealthcareApis.Commands
                     },
                     CosmosDbConfiguration = new ServiceCosmosDbConfigurationInfo()
                     {
+                        KeyVaultKeyUri = CosmosKeyVaultKeyUri ?? healthcareApisAccount.Properties.CosmosDbConfiguration.KeyVaultKeyUri,
                         OfferThroughput = CosmosOfferThroughput ?? healthcareApisAccount.Properties.CosmosDbConfiguration.OfferThroughput
                     },
                     CorsConfiguration = new ServiceCorsConfigurationInfo()
@@ -295,7 +419,7 @@ namespace Microsoft.Azure.Commands.HealthcareApis.Commands
 
             if (this.EnableManagedIdentity.ToBool() && healthcareApisAccount.Identity == null)
             {
-                servicesDescription.Identity = new Management.HealthcareApis.Models.ServicesResourceIdentity() { Type = "SystemAssigned" };
+                servicesDescription.Identity = new ServicesResourceIdentity() { Type = "SystemAssigned" };
             }
             else if (!this.DisableManagedIdentity.ToBool())
             {
@@ -303,7 +427,7 @@ namespace Microsoft.Azure.Commands.HealthcareApis.Commands
             }
             else
             {
-                servicesDescription.Identity = new Management.HealthcareApis.Models.ServicesResourceIdentity() { Type = "None" };
+                servicesDescription.Identity = new ServicesResourceIdentity() { Type = "None" };
             }
 
             return servicesDescription;
@@ -374,6 +498,7 @@ namespace Microsoft.Azure.Commands.HealthcareApis.Commands
                     },
                     CosmosDbConfiguration = new ServiceCosmosDbConfigurationInfo()
                     {
+                        KeyVaultKeyUri = InputObject.CosmosDbKeyVaultKeyUri ?? healthcareApisAccount.Properties.CosmosDbConfiguration.KeyVaultKeyUri,
                         OfferThroughput = InputObject.CosmosDbOfferThroughput ?? healthcareApisAccount.Properties.CosmosDbConfiguration.OfferThroughput
                     },
                     CorsConfiguration = new ServiceCorsConfigurationInfo()
@@ -397,7 +522,7 @@ namespace Microsoft.Azure.Commands.HealthcareApis.Commands
 
             if (!String.IsNullOrEmpty(InputObject.IdentityType))
             {
-                servicesDescription.Identity = new Management.HealthcareApis.Models.ServicesResourceIdentity(InputObject.IdentityPrincipalId, InputObject.IdentityTenantId, InputObject.IdentityType);
+                servicesDescription.Identity = new ServicesResourceIdentity(InputObject.IdentityPrincipalId, InputObject.IdentityTenantId, InputObject.IdentityType);
             }
             else
             {