+++ src/HealthcareApis/HealthcareApis/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added support for customer managed keys
 
 ## Version 3.0.0
 * Updated the SDK version to 3.0.0